fix for SOA serial overflows, by Dan Bilik & co-worker
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@626 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/resolver.cc
@@ -483,7 +483,7 @@ int Resolver::getSoaSerial(const string &ip, const string &domain, uint32_t *ser
   if(parts.size()<3)
     return 0;
   
-  *serial=atoi(parts[2].c_str());
+  *serial=(uint32_t)atol(parts[2].c_str());
   return 1;
 }
